Fixed Notice Types
