+++ packages/js/email-editor/src/hooks/use-validation-notices.ts
@@ -4,23 +4,19 @@
 import { useCallback } from '@wordpress/element';
 import { dispatch, useSelect } from '@wordpress/data';
 import { store as noticesStore } from '@wordpress/notices';
+import type {
+	NoticeAction,
+	NoticeListProps,
+} from '@wordpress/components/build-types/notice/types';
+import type { WPNoticeAction } from '@wordpress/notices/build-types/store/actions';
 
-export type NoticeAction = {
-	label: string;
-	onClick?: React.MouseEventHandler< HTMLButtonElement >;
-};
-
-export type Notice = {
-	id?: string;
-	title?: string;
-	content?: string;
-	className?: string;
-	type?: string;
-	children?: React.ReactNode;
-	isDismissible?: boolean;
-	handleDismiss?: () => void;
-	actions?: NoticeAction[];
-};
+// Re-export the canonical shapes from `@wordpress/components` so callers of
+// `addValidationNotice` use the same discriminated-union `NoticeAction` that
+// `<NoticeList>` accepts. `Notice` is the single-element shape that
+// `<NoticeList>` renders, which lines up with the narrowed projection of the
+// notices store's return.
+export type { NoticeAction };
+export type Notice = NoticeListProps[ 'notices' ][ number ];
 
 export type ValidationNoticesData = {
 	notices: Notice[];
@@ -35,9 +31,15 @@ export type ValidationNoticesData = {
 
 export const useValidationNotices = (): ValidationNoticesData => {
 	const context = 'email-validation';
-	const notices = useSelect( ( mapSelect ) =>
-		mapSelect( noticesStore ).getNotices( context )
+	const storeNotices = useSelect(
+		( mapSelect ) => mapSelect( noticesStore ).getNotices( context ),
+		[]
 	);
+	// `WPNotice.status: string` and `WPNoticeAction.onClick: Function` are
+	// wider than `<NoticeList>`'s element shape. At runtime the notices store
+	// only ever emits values inside the narrower union; cast once here so
+	// downstream code stays fully typed.
+	const notices = storeNotices as unknown as Notice[];
 
 	return {
 		notices,
@@ -55,11 +57,22 @@ export const useValidationNotices = (): ValidationNoticesData => {
 			[ notices ]
 		),
 		addValidationNotice: useCallback(
-			( noticeId: string, message: string, actions = [] ): void => {
+			(
+				noticeId: string,
+				message: string,
+				actions: NoticeAction[] = []
+			): void => {
+				// The notices store's `WPNoticeAction` is the same shape as
+				// `NoticeAction` at runtime, but its JSDoc-derived type widens
+				// `onClick` to `Function` and makes `url` required (allowing
+				// `null`). Cast at the boundary so the store sees its expected
+				// shape.
+				const storeActions =
+					actions as unknown as WPNoticeAction[];
 				void dispatch( noticesStore ).createNotice( 'error', message, {
 					id: noticeId,
 					isDismissible: false,
-					actions,
+					actions: storeActions,
 					context,
 				} );
 			},
