+++ packages/js/email-editor/typings/monorepo-enhancements.d.ts
@@ -1,0 +1,5 @@
+import '@woocommerce/internal-ts-config/types/@wordpress/data';
+import '@woocommerce/internal-ts-config/types/@wordpress/core-data';
+import '@woocommerce/internal-ts-config/types/@wordpress/block-editor';
+import '@woocommerce/internal-ts-config/types/@wordpress/notices';
+import '@woocommerce/internal-ts-config/types/@wordpress/notices';
